мелкие правки  #v.0.9.9.rev.A.build.31

+++ core/src/ru/asupd/poop_ballon/Workers/Shaker.java
@@ -20,7 +20,7 @@ public class Shaker {
        baseX+= camera.position.x ;
        baseY+= camera.position.y ;
     }
-    public void setUP_intensity(int i){
+    public void setUP_intensity(float i){
         intensity=i;
     }
     public void shake(float duration_in){
